requirements.txt
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-bootstrap4==1.0.1
django-chartjs==1.5.0
django-jet==1.0.8
idna==2.8
jet==1.0
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
